pyproject.toml
[tool.poetry]
name = "ginza-examples"
version = "0.1.0"
description = "GiNZA examples"
authors = ["poyo46 <poyo4rock@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
ginza = "^4.0.0"
rich = "^6.0.0"
sumy = "^0.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
PyYAML = "^5.3.1"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
